backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
numpy>=1.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AWS SDKs
boto3>=1.34.0
botocore>=1.34.0
dynamodb-json>=1.3.0

# Vector Search
pinecone-client>=2.2.4

# LLM and RAG
langchain>=0.1.0
langchain-aws>=0.1.0

# Additional utilities
redis>=5.0.0
opensearch-py>=2.4.0
pydantic-settings>=2.0.0